🏷️rename: TemplateCard => TemplateLayout
