--- src/components/server/TemplateCard/hooks/useCardSize/useCardSize.type.ts
@@ -1,5 +1,0 @@
-import { CardSizeType } from '../../TemplateCard.type';
-
-export interface UseCardSize {
-  type: CardSizeType;
-}
