+++ src/components/server/TemplateLayout/hooks/useCardSize/useCardSize.type.ts
@@ -1,0 +1,5 @@
+import { CardSizeType } from '../../TemplateLayout.type';
+
+export interface UseCardSize {
+  type: CardSizeType;
+}
