+++ src/components/server/index.ts
@@ -9,7 +9,7 @@ export { default as Carousel } from './Carousel/Carousel';
 export { default as EditorView } from './EditorView/EditorView';
 export { default as CommonList } from './CommonList/CommonList';
 export { default as Banner } from './Banner/Banner';
-export { default as TemplateCard } from './TemplateCard/TemplateCard';
+export { default as TemplateLayout } from './TemplateLayout/TemplateLayout';
 export { default as PhoneMockup } from './PhoneMockUp/PhoneMockup';
 export { default as InvitationItem } from './InvitationItem/InvitationItem';
 export { default as LoginContainer } from './LoginContainer/LoginContainer';
