+++ src/stories/components/TemplateCard_Story/TemplateCard.stories.tsx
@@ -1,20 +1,20 @@
-import { TemplateCard } from '@/components/server';
+import { TemplateLayout } from '@/components/server';
 import { Meta, StoryObj } from '@storybook/react';
 
 /**
- *  ## TemplateCard Component
+ *  ## TemplateLayout Component
  *
  * PinkCotton 내부에서 사용되어지는 Template Card 입니다
  *
  *  ### Props
  *  - **size : **`small` `medium` `large` 를 전달시 일반 Template Card가 제공됩니다.
  *  - **children ? : **Card 내부에 전달할 요소를 전달받습니다.
- *  - **className ? : **TemplateCard의 layout을 스타일링 할 수 있습니다. - **tailwindCSS**
+ *  - **className ? : **TemplateLayout의 layout을 스타일링 할 수 있습니다. - **tailwindCSS**
  */
 
 const meta = {
-  title: 'Components/TemplateCard',
-  component: TemplateCard,
+  title: 'Components/TemplateLayout',
+  component: TemplateLayout,
   parameters: {
     layout: 'centered',
   },
@@ -33,7 +33,7 @@ const meta = {
     },
     className: {
       control: 'text',
-      description: 'TemplateCard의 layout을 스타일링 할 수 있습니다. - **tailwindCSS**',
+      description: 'TemplateLayout의 layout을 스타일링 할 수 있습니다. - **tailwindCSS**',
     },
   },
 
@@ -43,7 +43,7 @@ const meta = {
   },
 
   decorators: [
-    (TemplateCardComponent) => {
+    (TemplateLayoutComponent) => {
       return (
         <section
           style={{
@@ -54,12 +54,12 @@ const meta = {
             justifyContent: 'center',
           }}
         >
-          <TemplateCardComponent />
+          <TemplateLayoutComponent />
         </section>
       );
     },
   ],
-} satisfies Meta<typeof TemplateCard>;
+} satisfies Meta<typeof TemplateLayout>;
 
 export default meta;
